+++ Library/Aphid/UI.js
@@ -27,7 +27,6 @@ Aphid.UI = {};
 //= require "UI/ImageView"
 //= require "UI/PopoverView"
 //= require "UI/LoadingIndicatorView"
-//= require "UI/LoadingIndicator"
 //= require "UI/AlertView"
 
 // List View -----------------------------------------------------------------